backend/requirements.txt
amqp==5.3.1
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.4
celery==5.6.3
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.7
daphne==4.2.1
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
exceptiongroup==1.3.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
kombu==5.6.2
msgpack==1.1.2
packaging==26.0
pillow==11.3.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.12.1
pyOpenSSL==26.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
redis==7.0.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.5
tomli==2.4.1
Twisted==25.5.0
txaio==23.6.1
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.0.1
